// src/models/accounts/account.ts
export enum AccountType {
  Checking = 'Checking',
  Savings = 'Savings',
  Credit = 'Credit',
  Investment = 'Investment',
  PersonalLoan = 'Personal Loan',
  Mortgage = 'Mortgage',
  AutoLoan = 'Auto Loan',
  BusinessChecking = 'Business Checking',
  BusinessSavings = 'Business Savings',
  BusinessCredit = 'Business Credit',
  Custodial = 'Custodial',
}